+++ dqops/src/integration-test/java/com/dqops/spark/sensors/column/strings/SparkColumnStringsStringMatchRegexPercentSensorParametersSpecIntegrationTest.java
@@ -66,7 +66,7 @@ public class SparkColumnStringsStringMatchRegexPercentSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(50.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(50.0d, resultTable.column(0).get(0));
     }
 
     @Test
@@ -80,7 +80,7 @@ public class SparkColumnStringsStringMatchRegexPercentSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(50.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(50.0d, resultTable.column(0).get(0));
     }
 
     @Test
@@ -94,7 +94,7 @@ public class SparkColumnStringsStringMatchRegexPercentSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(50.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(50.0d, resultTable.column(0).get(0));
     }
 
     @Test
@@ -108,7 +108,7 @@ public class SparkColumnStringsStringMatchRegexPercentSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(25, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 83.333, (float)resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 83.333, (double)resultTable.column(0).get(0), 0.001);
     }
 
     @Test
@@ -123,6 +123,6 @@ public class SparkColumnStringsStringMatchRegexPercentSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(50.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(50.0d, resultTable.column(0).get(0));
     }
 }
